Fix str diff algorithm
There was previously a bug where the optimal solution would not be found based on the order of the given inputs. Now the algorithm is run with inputs with inputs first_str, second_str as well as second_str, first_str as a work-around of this bug. This adds unnecessary computation but is a simple and readable fix.