Finalise packaging config
Restructured project to remove src/ directory because it is unnecessary at this small size and was causing problems with the sentinel singleton because it was being imported with two different file paths within diff.py and test_diff.py. This meant Python did not cache the value and was creating two separate sentinel objects in memory and causing failed assertions in the tests.